+++ package.json
@@ -20,11 +20,12 @@
     "clean": "gulp clean --color",
     "lint": "gulp tslint --color",
     "make": "gulp make --color",
-    "test": "gulp test --color"
+    "test": "jest --runInBand --colors"
   },
   "devDependencies": {
     "@angular/common": "~4.3.0",
     "@angular/compiler": "~4.3.0",
+    "@angular/compiler-cli": "~4.3.0",
     "@angular/core": "~4.3.0",
     "@angular/platform-browser": "~4.3.0",
     "@angular/platform-browser-dynamic": "~4.3.0",
@@ -32,7 +33,9 @@
     "core-js": "^2.5.0",
     "rxjs": "^5.4.3",
     "zone.js": "^0.8.17",
-    "@types/jasmine": "^2.5.54",
+    "lodash": "^4.17.4",
+    "@types/lodash": "4.14.55",
+    "@types/jest": "^20.0.8",
     "gulp": "gulpjs/gulp#4.0",
     "gulp-load-plugins": "^1.5.0",
     "rimraf": "^2.6.1",
@@ -41,20 +44,28 @@
     "webpack": "^3.5.5",
     "tslint-loader": "^3.5.3",
     "awesome-typescript-loader": "^3.2.3",
-    "jasmine-core": "^2.8.0",
-    "karma": "^1.7.1",
-    "karma-jasmine": "^1.1.0",
-    "karma-coverage": "^1.1.1",
-    "karma-mocha-reporter": "^2.2.4",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^2.0.4",
-    "karma-chrome-launcher": "^2.2.0",
-    "source-map-loader": "^0.2.1",
-    "istanbul-instrumenter-loader": "^3.0.0",
+    "jest": "^20.0.4",
+    "jest-preset-angular": "^2.0.5",
     "codelyzer": "^3.1.2",
     "tslint": "~5.7.0",
     "angular-tslint-rules": "^1.0.4",
     "typescript": "~2.3.4"
+  },
+  "jest": {
+    "preset": "jest-preset-angular",
+    "setupTestFrameworkScriptFile": "./config/jest.setup.ts",
+    "globals": {
+      "ts-jest": {
+        "tsConfigFile": "./tsconfig.json"
+      },
+      "__TRANSFORM_HTML__": true
+    },
+    "cache": false,
+    "noStackTrace": true,
+    "silent": true,
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "packages/@ngx-meta/core/src/**.ts"
+    ]
   }
 }
