--- packages/@ngx-meta/core/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "files": [
-    "./index.ts",
-    "./src/models/meta-settings.ts",
-    "./src/models/page-title-positioning.ts",
-    "./src/meta.guard.ts",
-    "./src/meta.loader.ts",
-    "./src/meta.service.ts",
-    "./src/util.ts"
-  ],
-  "angularCompilerOptions": {
-    "strictMetadataEmit": true,
-    "skipTemplateCodegen": true
-  }
-}
